backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==5.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.2
attrs==25.4.0
bandit==1.8.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
black==25.11.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
curl_cffi==0.14.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exchange_calendars==4.11.3
fastapi==0.120.4
fastapi-mail==1.6.2
flake8==7.1.1
frozendict==2.4.7
frozenlist==1.8.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
korean-lunar-calendar==0.3.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas_market_calendars==5.2.3
passlib==1.7.4
pathspec==0.12.1
peewee==3.18.3
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pyluach==2.3.0
pyparsing==3.3.2
pytest==8.4.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.49.1
stevedore==5.6.0
toolz==1.1.0
twilio==9.8.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
